+++ pypubtrack/cli.py
@@ -3,18 +3,26 @@ import sys
 import os
 import click
 import shutil
+import datetime
+
+import pybliometrics
+from pybliometrics.scopus import AbstractRetrieval, ScopusSearch
+from pybliometrics.scopus import config as scopus_config
 
 from pykitopen import KitOpen
 from pykitopen.search import YearBatching
 from pykitopen.publication import Publication
 from pykitopen.config import DEFAULT
 
-from pypubtrack.util import (get_config_path,
+from pypubtrack.util import (out,
+                             get_config_path,
                              get_version,
                              check_installation,
                              get_installation_path,
                              init_installation,
-                             author_name_kitopen)
+                             author_name_kitopen,
+                             get_template,
+                             ScopusPublicationAdapter)
 from pypubtrack.config import Config
 from pypubtrack.pypubtrack import Pubtrack
 
@@ -91,12 +99,117 @@ def init(ctx, force):
     click.secho('Created new installation folder at "{}"'.format(folder_path), fg='green', bold=True)
 
 
-@click.command('update-scopus')
-def update_scopus():
-    click.secho('update scopus')
+@click.command('config', short_help='Edit the config for ufotest')
+@click.option('--editor', '-e', type=click.STRING, help='Specify the editor command to be used to open the config file')
+def config(editor):
+    """
+    Edit the configuration file for this project.
+    """
+    config_path = get_config_path()
+    click.edit(filename=config_path, editor=editor)
+
+    return 0
 
 
-@click.command('update-kitopen')
+@click.command('import-scopus', short_help='Imports publications from scopus for the authors of the pubtrack app')
+@click.option('--verbose', '-v', is_flag=True, help='Print additional console output')
+@click.option('--start', '-s', type=click.INT, default='2015', help='The year to start the SCOPUS query from')
+@click.pass_context
+def import_scopus(ctx, verbose, start):
+    """
+    Import scopus publication records for the authors of the pubtrack application.
+
+    This command will first fetch all the information about the authors, which are defined within the pubtrack app.
+    It uses the scopus author ID's of these authors to send requests to the scopus database. The publications of these
+    replies are then evaluated and posted into the pubtrack app.
+    """
+    # SETTING UP PUBTRACK WRAPPER
+    config = ctx.obj['config']
+    pubtrack = Pubtrack(config)
+
+    # SETTING UP SCOPUS WRAPPER
+    try:
+        pybliometrics.scopus.utils.create_config()
+    except FileExistsError:
+        pass
+    finally:
+        scopus_config['Authentication']['APIKey'] = config.get_scopus_key()
+
+    # FETCHING META AUTHOR INFORMATION FROM PUBTRACK
+    click.secho('Fetching author information from pubtrack.')
+    author_id_name_map = {}
+    meta_authors = pubtrack.meta_author.get()['results']
+    for meta_author in meta_authors:
+        for author in meta_author['authors']:
+            # "author_name_kitopen" returns a string with the authors name. This function essentially formats the name
+            # in a way so that it can be used in a query string for the KITOpen database.
+            full_name = '{} {}'.format(author['first_name'], author['last_name'])
+            scopus_id = author['scopus_id']
+            author_id_name_map[scopus_id] = full_name
+            out(verbose, ' > Adding author "{} ({})" to be processed'.format(full_name, scopus_id))
+
+    click.secho('==> Processing total of {} authors'.format(len(author_id_name_map)))
+
+    # QUERY SCOPUS DATABASE
+    click.secho('Querying scopus database for the publications of those authors.')
+    date_limit = datetime.datetime(year=start, month=1, day=1)
+    for author_id, author_name in author_id_name_map.items():
+        publication_count = 0
+        search = ScopusSearch(f'AU-ID ( {author_id} )')
+        out(verbose, ' | Query "AU-ID ( {} )"'.format(author_id))
+
+        for result in search.results:
+
+            # We'll only take publications, which have a DOI
+            if result.doi is None:
+                continue
+
+            # requesting the detailed information from the scopus database for the current publication from the search
+            # results
+            try:
+                abstract_retrieval = AbstractRetrieval(result.doi)
+            except Exception as e:
+                out(verbose, '   # Could not retrieve publication "{}"'.format(result.doi), fg='yellow')
+                continue
+
+            # If the publication is older than the date limit, it will be discarded
+            publication_date = datetime.datetime.strptime(abstract_retrieval.coverDate, '%Y-%m-%d')
+            if publication_date <= date_limit:
+                out(verbose, '   # Publication too old "{}"({})'.format(result.doi, publication_date), fg='yellow')
+                continue
+            else:
+                out(verbose, '   > Fetched publication "{}"'.format(result.doi))
+
+            adapter = ScopusPublicationAdapter(abstract_retrieval)
+            publication = adapter.get_publication()
+
+            # Filtering the authors according to the AUTHOR_LIMIT, which has been set.
+            # We cannot just use the first few authors however, we need to make sure that the author, from which we have
+            # this publication in the first place is in there. The rest just gets filled up...
+            authors = []
+            for author in publication['authors']:
+                if author['scopus_id'] in author_id_name_map.keys() or len(authors) < config.get_author_limit():
+                    authors.append(author)
+
+            publication['authors'] = authors
+
+            # Now we try to actually POST the publication to the pubtrack REST API
+            try:
+                pubtrack.import_publication(publication)
+                publication_count += 1
+                out(verbose, '   * Added to pubtrack: "{}"'.format(publication['title']), fg='green')
+            except Exception as e:
+                if str(e) == 'uuid':
+                    out(verbose, '   ! Error while posting to pubtrack: Already exists!', fg='red')
+                else:
+                    out(verbose, '   ! Error while posting to pubtrack: {}'.format(str(e)), fg='red')
+                continue
+
+        out(True, ' --> Total of {} publications imported from author {}'.format(publication_count, author_id),
+            fg='green', bold=True)
+
+
+@click.command('update-kitopen', short_help='Updates existing pubtrack publication records with KITOpen information')
 @click.option('--verbose', '-v', is_flag=True, help='Print additional console output')
 @click.option('--start', '-s', type=click.STRING, default='2015', help='The year to start the KITOpen query from')
 @click.pass_context
@@ -121,8 +234,7 @@ def update_kitopen(ctx, verbose, start):
             # in a way so that it can be used in a query string for the KITOpen database.
             name = author_name_kitopen(author['first_name'], author['last_name'])
             author_names.append(name)
-            if verbose:
-                click.secho(' * Adding author "{}" to query'.format(name))
+            out(verbose, ' > Adding author "{}" to query'.format(name))
 
     click.secho('==> Processing total of {} authors'.format(len(author_names)))
 
@@ -175,23 +287,52 @@ def update_kitopen(ctx, verbose, start):
                 # DOI.
                 pubtrack.publication.patch(pub['uuid'], patch=patch)
                 count_success += 1
-                if verbose:
-                    click.secho(' > updated publication {}'.format(publication.data['doi']))
+                out(verbose, ' > updated publication {}'.format(publication.data['doi']))
             except Exception as e:
-                if verbose:
-                    click.secho(' ! Warning updating publication {}'.format(publication.data['doi']), fg='yellow')
+                out(verbose, ' # Warning updating publication {}'.format(publication.data['doi']),
+                    fg='yellow')
             finally:
                 count_total += 1
 
-    click.secho('==> Updated {}/{} publications with KITOpen data'.format(count_success, count_total),
-                fg='green', bold=True)
+    out(True, '==> Updated {}/{} publications with KITOpen data'.format(count_success, count_total),
+        fg='green', bold=True)
 
     return 0
 
 
+@click.command('list-publications', short_help='Displays a list of publications, currently on the pubtrack app')
+@click.option('--verbose', '-v', is_flag=True, help='Print additional console output')
+@click.option('--uuid', '-u', type=click.STRING, help='Specify the UUID of an individual publication to display')
+@click.option('--doi', '-d', type=click.STRING, help='Specify the DOI of an individual publication to display')
+@click.pass_context
+def list_publications(ctx, verbose, uuid, doi):
+    # SETTING UP PUBTRACK WRAPPER
+    config = ctx.obj['config']
+    pubtrack = Pubtrack(config)
+
+    # FETCHING PUBLICATION DETAILS FROM PUBTRACK
+    if uuid:
+        publications = [pubtrack.publication.get(uuid)]
+    elif doi:
+        publications = [pubtrack.publication.get_by(doi=doi)]
+    else:
+        publications = pubtrack.publication.get()['results']
+
+    # DISPLAYING PUBLICATIONS TO THE USER
+    for publication in publications:
+        if verbose:
+            template = get_template('publication.j2')
+            info = template.render(publication=publication)
+            out(verbose, info)
+        else:
+            out(True, '{} ({})'.format(publication['title'], publication['doi']))
+
+
 cli.add_command(init)
-cli.add_command(update_scopus)
+cli.add_command(config)
+cli.add_command(import_scopus)
 cli.add_command(update_kitopen)
+cli.add_command(list_publications)
 
 
 if __name__ == "__main__":